tp promise async

+++ index.js
@@ -3,42 +3,41 @@ const fs = require('fs-extra');
 const file = 'pubs.json';
 const dir = './temp/';
 
-
-fs.ensureDir(dir)
+// on vérifie d'abord la présence du repertoire temp avant d'appeler d'autre promise
+fs.pathExists(dir)
+    .then(exist => {
+        if (exist) {
+            console.log("=== remove ");
+            return fs.remove(dir)
+        }
+        return;
+    })
     .then(() => {
-        console.log('=== Création Temp ===')
+        console.log("=== create dir & file ");
+        return fs.ensureDir(dir) && fs.copy('pubs.json', dir + 'pubs.json')
+    })
+
+
+    .then(() => {
+        console.log("== lecture json");
+
+        return fs.readJson('pubs.json').then((pubs) => {
+            console.log(pubs.name)
+        })
+
+    })
+    .then(() => {
+        console.log("=== watcher ");
+        return fs.watchFile('pubs.json', (curr, prev) => {
+
+            console.log(`the current mtime is: ${curr.mtime}`);
+            console.log(`the previous mtime was: ${prev.mtime}`);
+        })
     })
     .catch(err => {
         console.error(err)
     })
 
 
-fs.remove(dir)
-    .then(() => {
-        console.log('=== Supression Temp ===')
-    })
-    .catch(err => {
-        console.error(err)
-    })
 
 
-fs.pathExists(file)
-    .then(exists => console.log(exists)) 
-
-
-/*
-fs.readFile('pubs.json', 'utf-8', function (e, data) {
-    if (e) {
-        console.log("log = " + e);
-    }
-    console.log("file = " + data);
-
-});
-
-fs.watchFile('pubs.json', (curr, prev) => {
-
-    console.log(`the current mtime is: ${curr.mtime}`);
-    console.log(`the previous mtime was: ${prev.mtime}`);
-});
-
-*/
